+++ src/command/commands/song.ts
@@ -6,7 +6,7 @@ const command: Command = {
   help: "Information about the song being played if it's part of a queue.",
   aliases: ['music'],
   platforms: [Platform.Twitch],
-  run: async (ctx, { services }) => {
+  run: async (ctx, { services, commandState }) => {
     if (!CONFIG.get().spotify?.enabled) {
       console.log('Spotify is not enabled for !song.');
       return false;
@@ -16,10 +16,14 @@ const command: Command = {
       ctx.reply(ctx, "I couldn't find the current song.");
       return false;
     }
-    ctx.reply(
-      ctx,
-      `Currently playing: ${services.spotifyService.getTrackInfoFromTrack(track)}.`
-    );
+    const trackDetails = services.spotifyService.getTrackInfoFromTrack(track);
+    const requester = commandState.spotify.requests[trackDetails];
+
+    let reply = `Currently playing: ${trackDetails}.`;
+    if (requester) {
+      reply += ` Requested by ${requester}.`;
+    }
+    ctx.reply(ctx, reply);
     return true;
   }
 };