+++ src/command/commands/songrequest.ts
@@ -6,7 +6,7 @@ const command: Command = {
   help: 'Request a song to be played on stream.',
   aliases: ['sr'],
   platforms: [Platform.Twitch],
-  run: async (ctx, { services }) => {
+  run: async (ctx, { services, commandState }) => {
     if (!CONFIG.get().spotify?.enabled) {
       console.log('Spotify is not enabled for !songrequest.');
       return false;
@@ -17,7 +17,16 @@ const command: Command = {
       return false;
     }
 
-    const track = await services.spotifyService.getTrackFromSearch(query);
+    let track;
+
+    if (query.includes('spotify.com/track/')) {
+      // Get the id from url that looks like https://open.spotify.com/track/0dymahiZQrRBzYTRg9QK9i?si=fb0b49338d6a4f7c
+      const id = query.split('track/')[1].split('?')[0];
+      track = await services.spotifyService.getTrackById(id);
+    } else {
+      track = await services.spotifyService.getTrackFromSearch(query);
+    }
+
     if (!track) {
       ctx.reply(ctx, "I couldn't find that song on Spotify.");
       return false;
@@ -35,10 +44,10 @@ const command: Command = {
     }
     const added = await services.spotifyService.addTrackToQueue(track.uri);
     if (added) {
-      ctx.reply(
-        ctx,
-        `${services.spotifyService.getTrackInfoFromTrack(track)} added to queue.`
-      );
+      const user = ctx.onBehalfOf || ctx.displayName;
+      const trackDetails = services.spotifyService.getTrackInfoFromTrack(track);
+      commandState.spotify.requests[trackDetails] = user;
+      ctx.reply(ctx, `${trackDetails} added to queue.`);
       return true;
     } else {
       ctx.reply(ctx, "I couldn't add that song to the queue.");