+++ src/twitch/twitch.eventsub.ts
@@ -61,13 +61,17 @@ export class TwitchEventSub {
     }
 
     if (messageType === 'session_reconnect') {
-      this.eventSubCreateConnection(parsedData.payload.session.reconnect_url);
+      void this.eventSubCreateConnection(
+        parsedData.payload.session.reconnect_url
+      );
       return;
     }
 
     if (messageType === 'revocation') {
       this.eventSubConnection.close();
-      this.eventSubCreateConnection(parsedData.payload.session.reconnect_url);
+      void this.eventSubCreateConnection(
+        parsedData.payload.session.reconnect_url
+      );
     }
 
     if (messageType === 'notification') {
@@ -82,40 +86,40 @@ export class TwitchEventSub {
       try {
         switch (subType) {
           case 'channel.follow':
-            await this.onFollow(parsedData);
+            await this.onFollow(parsedData.payload.event);
             break;
 
           case 'stream.online':
-            await this.onStreamOnline(parsedData);
+            await this.onStreamOnline(parsedData.payload.event);
             break;
 
           case 'channel.hype_train.begin':
-            await this.onTrain(parsedData);
+            await this.onTrain(parsedData.payload.event);
             break;
 
           case 'channel.hype_train.progress':
-            await this.onTrainProgress(parsedData);
+            await this.onTrainProgress(parsedData.payload.event);
             break;
 
           case 'channel.raid':
-            await this.onRaid(parsedData);
+            await this.onRaid(parsedData.payload.event);
             break;
 
           case 'channel.cheer':
-            await this.onCheer(parsedData);
+            await this.onCheer(parsedData.payload.event);
             break;
 
           case 'channel.subscribe':
             // NOTE: This doesn't include resubscribes
-            await this.onSubscribe(parsedData);
+            await this.onSubscribe(parsedData.payload.event);
             break;
 
           case 'channel.subscription.gift':
-            await this.onSubscriptionGift(parsedData);
+            await this.onSubscriptionGift(parsedData.payload.event);
             break;
 
           case 'channel.chat.message':
-            await this.onChatMessage(parsedData);
+            await this.onChatMessage(parsedData.payload.event);
             break;
 
           default:
@@ -195,105 +199,70 @@ export class TwitchEventSub {
     }
   }
 
-  async onChatMessage(data: {
-    payload: {
-      event: {
-        chatter_user_name: string;
-        chatter_user_login: string;
-        message: {
-          text: string;
-        };
-        badges: { set_id: string }[];
-        message_type: string;
-        channel_points_custom_reward_id: string | null;
-      };
-    };
-  }) {
+  async onChatMessage(data: OnChatMessageEvent) {
     const obj: OnMessageHandlerInput = {
-      message: data.payload.event.message.text,
-      userLogin: data.payload.event.chatter_user_login,
-      displayName: data.payload.event.chatter_user_name,
-      isMod: data.payload.event.badges.some(
-        (badge) => badge.set_id === 'moderator'
-      ),
-      isSub: data.payload.event.badges.some(
-        (badge) => badge.set_id === 'subscriber'
+      message: data.message.text,
+      userLogin: data.chatter_user_login,
+      displayName: data.chatter_user_name,
+      isMod: data.badges.some((badge) => badge.set_id === 'moderator'),
+      isSub: data.badges.some((badge) => badge.set_id === 'subscriber'),
+      isVip: data.badges.some((badge) => badge.set_id === 'vip'),
+      isFounder: data.badges.some((badge) => badge.set_id === 'founder'),
+      isHypeTrainConductor: data.badges.some(
+        (badge) => badge.set_id === 'hype-train'
       ),
       isOwner:
-        data.payload.event.chatter_user_login ===
+        data.chatter_user_login ===
         CONFIG.get().twitch.ownerUsername.toLowerCase(),
-      channelPointsCustomRewardId:
-        data.payload.event.channel_points_custom_reward_id
+      channelPointsCustomRewardId: data.channel_points_custom_reward_id
     };
 
     void this.twitchService.onMessageHandler(obj);
   }
 
-  async onCheer(data: {
-    payload: {
-      event: {
-        message: string;
-        bits: string;
-        user_name: string;
-        is_anonymous: boolean;
-      };
-    };
-  }) {
+  async onCheer(data: OnCheerEvent) {
     const obj = {
-      message: data.payload.event.message,
-      bits: parseInt(data.payload.event.bits) || 0,
-      user: data.payload.event.is_anonymous
-        ? 'Anonymous'
-        : data.payload.event.user_name
+      message: data.message,
+      bits: data.bits || 0,
+      user: data.is_anonymous ? 'Anonymous' : data.user_name
     };
 
     void this.twitchService.ownerRunCommand(`!onbits ${JSON.stringify(obj)}`);
   }
 
-  async onFollow(data: { payload: { event: { user_login: string } } }) {
-    void this.twitchService.helixGetTwitchUserInfo(
-      data.payload.event.user_login,
-      true
-    );
+  async onFollow(data: OnFollowEvent) {
+    void this.twitchService.helixGetTwitchUserInfo(data.user_login, true);
   }
 
-  async onSubscribe(data: {
-    payload: { event: { user_name: string; is_gift?: string } };
-  }) {
-    if (data.payload.event.is_gift) {
+  async onSubscribe(data: OnSubscribeEvent) {
+    if (data.is_gift) {
       // This will be handled by the gift sub event
       return;
     }
     const obj = {
-      username: data.payload.event.user_name
+      username: data.user_name
     };
     void this.twitchService.ownerRunCommand(`!onsubs ${JSON.stringify(obj)}`);
   }
 
-  async onSubscriptionGift(data: {
-    payload: { event: { user_name: string; is_anonymous: string } };
-  }) {
+  async onSubscriptionGift(data: OnSubscriptionGiftEvent) {
     const obj = {
-      username: data.payload.event.is_anonymous
-        ? 'Anonymous'
-        : data.payload.event.user_name,
+      username: data.is_anonymous ? 'Anonymous' : data.user_name,
       isGifter: true
     };
     void this.twitchService.ownerRunCommand(`!onsubs ${JSON.stringify(obj)}`);
   }
 
-  async onRaid(data: {
-    payload: { event: { from_broadcaster_user_login: string } };
-  }) {
+  async onRaid(data: OnRaidEvent) {
     void this.twitchService.ownerRunCommand(
-      `!onraids ${data.payload.event.from_broadcaster_user_login}`
+      `!onraids ${data.from_broadcaster_user_login}`
     );
     void this.twitchService.ownerRunCommand(
-      `!so ${data.payload.event.from_broadcaster_user_login}`
+      `!so ${data.from_broadcaster_user_login}`
     );
   }
 
-  async onTrain(_data: unknown) {
+  async onTrain(_data: OnHypeTrainBeginEvent) {
     this.currentHypeTrainLevel = 1;
     this.twitchService.botSpeak(
       `Hype train level 1! Come on the train, choo choo ride it!!`
@@ -301,8 +270,8 @@ export class TwitchEventSub {
     void this.twitchService.ownerRunCommand('!train');
   }
 
-  async onTrainProgress(data: { payload: { event: { level: number } } }) {
-    const level = data.payload.event.level;
+  async onTrainProgress(data: OnHypeTrainProgressEvent) {
+    const level = data.level;
     if (level > this.currentHypeTrainLevel) {
       this.currentHypeTrainLevel = level;
       this.twitchService.botSpeak(`Hype train level ${level}! LET'S GOOOOOOO!`);
@@ -310,7 +279,7 @@ export class TwitchEventSub {
     }
   }
 
-  async onStreamOnline(_data: unknown) {
+  async onStreamOnline(_data: OnStreamOnlineEvent) {
     this.logger.log('Stream is online!');
   }
 }